extract DB lock into a store type
