+++ pkg/setting/setting.go
@@ -546,6 +546,9 @@ type Cfg struct {
 
 	// Feature Management Settings
 	FeatureManagement FeatureMgmtSettings
+
+	// ServerLock
+	ServerLock ServerLockSettings
 }
 
 // AddChangePasswordLink returns if login form is disabled or not since
@@ -1222,18 +1225,8 @@ func (cfg *Cfg) Load(args CommandLineArgs) error {
 	enterprise := iniFile.Section("enterprise")
 	cfg.EnterpriseLicensePath = valueAsString(enterprise, "license_path", filepath.Join(cfg.DataPath, "license.jwt"))
 
-	cacheServer := iniFile.Section("remote_cache")
-	dbName := valueAsString(cacheServer, "type", "database")
-	connStr := valueAsString(cacheServer, "connstr", "")
-	prefix := valueAsString(cacheServer, "prefix", "")
-	encryption := cacheServer.Key("encryption").MustBool(false)
-
-	cfg.RemoteCacheOptions = &RemoteCacheOptions{
-		Name:       dbName,
-		ConnStr:    connStr,
-		Prefix:     prefix,
-		Encryption: encryption,
-	}
+	cfg.RemoteCacheOptions = readRemoteCacheSettings(iniFile)
+	cfg.ServerLock = readServerLockSettings(iniFile)
 
 	geomapSection := iniFile.Section("geomap")
 	basemapJSON := valueAsString(geomapSection, "default_baselayer_config", "")
@@ -1271,13 +1264,6 @@ func (cfg *Cfg) Load(args CommandLineArgs) error {
 
 func valueAsString(section *ini.Section, keyName string, defaultValue string) string {
 	return section.Key(keyName).MustString(defaultValue)
-}
-
-type RemoteCacheOptions struct {
-	Name       string
-	ConnStr    string
-	Prefix     string
-	Encryption bool
 }
 
 func (cfg *Cfg) readSAMLConfig() {
